refactor feed request serialization
